added e2e tests

+++ server/src/routes/items/delete-todo-item.ts
@@ -4,7 +4,13 @@ import { TodoItem } from '../../model/todo-item';
 
 // get all items from database
 export const deleteTodoItem = (req: Request, res: Response) => {
+  console.log("req", req.body)
   let item: TodoItem = {id: req.body.id, todo: req.body.todo, completed: req.body.completed}
+  console.log('look here', item.id)
+  if(item.id === undefined) {
+    return res.sendStatus(400)
+    
+  }
   try {
     deleteItem(item)
     .then((result) => {
